+++ src/app/admin/brands/components/brandsTable/columns.tsx
@@ -65,7 +65,7 @@ export const getBrandsTableColumns = ({
             <Button
               variant={null}
               size="icon"
-              className="size-8 mr-2 cursor-pointer"
+              className="size-8 mr-2"
               onClick={() => onEdit(row.original)}
             >
               <SquarePen />
@@ -77,7 +77,7 @@ export const getBrandsTableColumns = ({
               <Button
                 variant={null}
                 size="icon"
-                className="size-8 cursor-pointer text-red-500"
+                className="size-8 text-red-500"
                 onClick={() => onDelete(row.original)}
               >
                 <Trash />
